refactor : Folder Name Renameing pado-server -> api-server
