+++ api-server/src/main/java/org/pado/api/HelloController.java
@@ -1,4 +1,4 @@
-package org.pado.server;
+package org.pado.api;
 
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RestController;
