+++ api-server/src/main/java/org/pado/api/PadoServerApplication.java
@@ -1,4 +1,4 @@
-package org.pado.server;
+package org.pado.api;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
